[branchLayout] Move BranchLayoutEntry to api
